requirements.txt
openai==1.9.0
pdfplumber==0.11.7
python-dotenv==1.0.1
google-cloud-vision==3.10.2
Pillow==11.3.0
pytesseract==0.3.13